Migrate to autoprobes.py
Also give up on global TYPES and use dependency injection instead.

Still WIP, so [ci skip]
